1 - Fixing svg skill sizes 2 - Responsive Navbar 3 - Responsive Hero 4 - Responsive Projects 5 - Hamburguer menu icon 6 - Changing project cards to rows
